+++ packages/backend/src/plugins/kubernetes.ts
@@ -15,12 +15,11 @@
  */
 
 import { createRouter } from '@backstage/plugin-kubernetes-backend';
-import { Router } from 'express';
 import { PluginEnvironment } from '../types';
 
 export default async function createPlugin({
   logger,
   config,
-}: PluginEnvironment): Promise<Router> {
+}: PluginEnvironment) {
   return await createRouter({ logger, config });
 }