+++ packages/theme/src/baseTheme.ts
@@ -257,6 +257,9 @@ export function createThemeOverrides(theme: BackstageTheme): Overrides {
         // etc) end up at the bottom of the card instead of just below the body
         // contents.
         flexGrow: 1,
+        '&:last-child': {
+          paddingBottom: undefined,
+        },
       },
     },
     MuiCardActions: {