+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/publish/githubPullRequest.ts
@@ -87,7 +87,9 @@ export const defaultClientFactory = async ({
   }
 
   const { token } = await credentialsProvider.getCredentials({
-    url: `${host}/${encodeURIComponent(owner)}/${encodeURIComponent(repo)}`,
+    url: `https://${host}/${encodeURIComponent(owner)}/${encodeURIComponent(
+      repo,
+    )}`,
   });
 
   if (!token) {