+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage/theme
 
+## 0.2.7
+
+### Patch Changes
+
+- 7b8272fb7: Remove extra bottom padding in InfoCard content
+
 ## 0.2.6
 
 ### Patch Changes