+++ plugins/rollbar-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-rollbar-backend
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [e0bfd3d44]
+- Updated dependencies [38ca05168]
+- Updated dependencies [d8b81fd28]
+  - @backstage/backend-common@0.7.0
+  - @backstage/config@0.1.5
+
 ## 0.1.9
 
 ### Patch Changes